Create the initial release skeleton
This was generated with a simple `bosh init release --git`. I manually added
some `.gitkeep` files to some of the empty directories so that it would be
easier to see which directories were created at each step.

A fake `final.yml` has been added too. This is required for some BOSH commands
such as creating a release. However, the values of the file are not checked.