Add Redis source code to release
Since we're going to be relying on the `unstable` branch of Redis it's best to
use the source code rather than uploading a release tarball to the blobstore.
The source was added using a git submodule with the command:

    $ git submodule add https://github.com/antirez/redis src/redis

The `unstable` branch was then checked out using the commands:

    $ cd src/redis
    $ git checkout unstable

We're using this branch because it contains the Redis Sentinel work that we're
going to be using later in order to manage automatic failover leader promotion,
monitoring, and error notifications.