Add Redis package
This commit adds the Redis package to the release. A package is used to define
how to compile source code or blobstore artefacts into a form that is usable
in the jobs of the release. A package may also depend on other packages if they
are required for its compilation.

The basic structure of the package was generated using the command:

    $ bosh generate package redis

The `redis` package begins with its specification which is defined in the
`packages/redis/spec` file. This file gives the package a name, any other
package dependencies that are required to compile this package, and a list of
source code or blobs that are required for compilation. These artefacts can
either be from the `blobs` or `src` directories of the release.

The other file in the package is the compilation script which is found at
`packages/redis/packaging`. This file tells BOSH how to take the artefacts
described above and turn them into something that another package or another
job may use. This file begins with setting some Bash settings in order to make
any errors during the compilation phase of this package easier to debug. We
then change into the `redis` directory which we chose for inclusion in the
specification file and the compile Redis (the origin of this command is the
Redis README). We set the `PREFIX` to be `$BOSH_INSTALL_TARGET` for this
command. The `$BOSH_INSTALL_TARGET` is the directory that BOSH expects all of
the compiled artefacts to be in at the end of the compilation. It will then zip
up the contents of this directory before saving the compiled package in its
blobstore for future use.