Add the Redis job
This is nearly the simplest (useful) job possible. It was generated initially
with the command:

    $ bosh generate job redis

The job contains a `monit` file, a `spec` file, and a template (`redis_ctl`)
that we'll combine with Monit to start and manage the job.

The `monit` file contains a Monit job configuration file that tells Monit the
name (that Monit will use), PID file, and start and stop commands for the Redis
job.

The `spec` file contains the specification for the job. This is similar to the
spec file for the package that we saw before. It is a YAML file that contains
the name of the job, the list of templates and where the populated versions
should be placed. By putting the `redis_ctl` file in the `bin/` directory BOSH
will automatically give executable permissions to the file. The last item in
this file are the packages that are required in order to run this job.

The last file is the `redis_ctl` file that is called by Monit when the job is
started and stopped. This file sets up the logging, PID file and starting of
the `redis-server` executable.