Scale our deployment to double the size
This commit shows the ease of scaling up a deployment using BOSH. We simply
change the deployment manifest and run `bosh deploy` again. The BOSH director
will realise that the deployment size needs to change and carry out the
appropriate scaling.

In this deployment there are three different values we need to change in order
to scale the deployment.

* The size of the resource pool is increased in order to allow more instances
  to be created inside it.

* The number of instances of the `redis` job is also increased. These instances
  will fill the newly enlarged resource pool.

* Another static IP is added to the job for the newly created instance.

This method will be applicable to any magnitude of scaling. In this commit we
only added a single instance. However, we could have just as easily changed
this number by 10, 100, or even a 1000 and BOSH will do the right thing.