Allow the configuration of the Redis password
BOSH injects configuration onto the virtual machines before it starts the jobs
on them. This commit shows a common method of making a configuration value of
a job be configurable by a release deployer. In this case we will be
configuring the authentication password for the Redis server. There's no real
order to these steps as they all depend on each other in some way.

The `spec` file of the Redis job must be modified by adding entry for the
templating of the Redis configuration file that we are going to add and adding
an entry for the new configurable property we would like people to be able to
configure.

Next, we add the example Redis configuration file from the Redis source code to
the path that we specified in the `spec` (`redis.conf.erb`). This has one
change on lines 310-312 which, if the property `redis.password` is set will
change the configuration of the Redis server to require a password to
authenticate. The `if_p` helper evaluates the block if the property name it is
given as an argument is present. The argument passed to the block is the value
of the property.

The startup script for the job, `redis_ctl`, must also be changed to use the
new configuration file. After BOSH has laid all of the templated files out on
disk the configuration file that we are looking for will be at the path
`/var/vcap/jobs/redis/config/redis.conf`. The last directory and filename in
that path is the same one that we specified in the `spec` file above.

The last change is to add the property to our deployment manifest. This is
where the value of our property originates before being distributed to our
deployment with our next `bosh deploy`. In the deployment manifest you can see
how the property in the YAML file are translated, through the `spec`, to real
values on machines.