Add Master-Slave replication to release
Now things are getting interesting. We've taken our deployment of many distinct
Redis instances and converted them into something that starting to look useful.
After this commit your Redis cluster will replicating from the master out to
N slaves.

We must change the `spec` file for the `redis` job and add 3 new properties:

* `redis.replication.enabled`
  This property must be set to true on each job instance, master or slave, that
  is part of a cluster. It's set to `false` by default to provide a sensible
  initial value.

* `redis.replication.master.enabled`
  This property is set to true on any job instance that should be the master of
  a Redis cluster. It too has a sensible default of false.

* `redis.replication.master.address`
  This property should be set on each slave to tell it the IP address that
  should be used to reach the master node.

Using these new properties we are able to add additional logic to the ERB in
the Redis configuration file to support clustering. We check to see if
replication has been enabled, if we have been given the IP address of the
master, and if this is not the master of the cluster. If all of these
predicates are true then we set the slave to follow a master and authenticate
with it if required.

Finally, we add our new configuration possibilities to the manifest. We now
need two jobs in our deployment: a master and a slave. However, because we did
some more work up front in our templating we are able to use the same
job-template for both of them: `redis`. Properties can either be specified
globally in the `properties` block at the bottom of the deployment manifest or
per-job. We configure the master node at the job level to be the master but
configure the cluster globally using our new properties.