Add HAProxy Package
Our users are getting fed up with having to connect to different IP addresses
to reach the same dataset. It's fragile and confusing. It would be much better
to put some sort of load balancing or routing layer into our release. Let's use
the venerable HAProxy to accomplish this.

We add the HAProxy source code tarball to our BOSH blobs the same as the Redis
package. In the `spec` we specify the name of our job, the fact that it has no
dependencies, and tell it to use the HAProxy blob that we added just now.

In our `packaging` script we step through the motions of compiling the source
into its binary form. All of these steps were taken directly from the HAProxy
documentation.