Add HAProxy Job
Now that we have our package all ready to go let's write the job that sits on
top of it.

This should all look familiar by now. We set up a job and fill in the blanks in
a very similar way to the Redis job. There is a new property called
`haproxy.backend.servers` which enumerates the backend Redis servers that we
are going to be proxying to.