Allow Redis port number to be configured
We have the same constant repeated all over our templates: the Redis port
number. Similarly to if we were writing code let's extract this number to
a variable and pass it in. We introduce the variable in all of the `spec` files
that care about the value and add the ERB required to inject it. This time we
add a default (the default Redis port number) to the property so that it does
not need to be set in the manifest.

Now we can set the port and change it at a whim as easily as the Redis
password.